% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_ai.R
\name{call_ai}
\alias{call_ai}
\title{Call AI}
\usage{
call_ai(input)
}
\arguments{
\item{input}{The prompt to be submitted to Gemma 3 12B}
}
\value{
Gemma 3 12B's response to the submitted prompt
}
\description{
Sends a prompt to the Google's Gemma 3 12B model.
Requires a free OpenRouter API, which can be obtained on \href{https://openrouter.ai/}{the OpenRouter website}.
}
\examples{
\dontrun{
call_ai(
input = "How can I learn swimming?"
)
}
}
\keyword{internal}
