% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api.R
\name{set_api}
\alias{set_api}
\title{Set API}
\usage{
set_api(api)
}
\arguments{
\item{api}{The API to be set. Has to be entered as a character string.}
}
\value{
Nothing
}
\description{
Sets the free OpenRouter API, which is necessary for the \code{\link{check_jailbreak}}
function and can be obtained on \href{https://openrouter.ai/}{the OpenRouter website}.
}
\examples{
\dontrun{
set_api(
api = "your_openrouter_api"
)
}
}
